Keep formatter range tests within ratchet
Move the range-boundary regression into its own formatter test module so the constrained formatter_tests file remains below the source-size ceiling.
